requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
awscli==1.16.157
backcall==0.1.0
bleach==3.1.0
boto3==1.9.134
botocore==1.12.147
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.6.1
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
findspark==1.3.0
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
idna==2.8
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==7.0.0
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.11.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pycparser==2.19
Pygments==2.4.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyrsistent==0.14.11
PySocks==1.6.8
pytest==4.5.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
pyzmq==18.0.0
qtconsole==4.4.4
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.0
scikit-learn==0.20.3
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
sql-magic==0.0.4
SQLAlchemy==1.3.3
sqlparse==0.3.0
statsmodels==0.9.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
xgboost==0.82
